# src/main/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-service-financial-aid
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
  - url: http://localhost:59933
    description: Generated server url
paths:
  /{municipalityId}/financial-aid:
    get:
      tags:
        - financial-aid-resource
      operationId: getFinancialAidBasis
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: personalNumber
          in: query
          required: true
          schema:
            type: string
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties: { }
  /{municipalityId}/financial-aid/connection:
    get:
      tags:
        - financial-aid-resource
      operationId: testConnection
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties: { }
  /api-docs:
    get:
      tags:
        - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  securitySchemes: { }